refactor distiller pipeline into different files
