+++ operators/distiller-pipeline/tests/conftest.py
@@ -1,7 +1,8 @@
 import numpy as np
 import pytest
 
-from distiller_pipeline import FrameAccumulator, FrameHeader
+from distiller_pipeline.accumulator import FrameAccumulator
+from distiller_pipeline.models import FrameHeader
 
 
 @pytest.fixture
